Update data_visualazer module
Now to plot we use a np.ndarray instead of a gb.MVar such that we can
also plot results stored in a np.ndarray (as in MSPP-PD(NQP) notebook).

Removed the annotations in the plot for the agents' source and terminus
for clearer plots.

Add module description for clarity