Add small fixes
In extra we are now sure to generate agents with random seed equal to 0
in the 2nd section even if we do not run the code from the first one.
Now we also print the distance convered and the penalty incurred for the
real case example.

In Model_practability the plot is now more understandable